be/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
googlemaps==4.6.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
